SentientShop-backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary
python-dotenv
djangorestframework-simplejwt==5.3.1
stripe==8.4.0